refactor : PathFinder 의 위치를 적절한 위치로 이동
- asis : path.finder
- tobe : path.domain
