+++ src/main/java/nextstep/subway/path/application/PathService.java
@@ -4,8 +4,8 @@ import java.util.List;
 import nextstep.subway.line.application.LineService;
 import nextstep.subway.line.domain.Line;
 import nextstep.subway.path.dto.PathResponse;
-import nextstep.subway.path.finder.DijkstraPathFinder;
-import nextstep.subway.path.finder.PathFinder;
+import nextstep.subway.path.domain.DijkstraPathFinder;
+import nextstep.subway.path.domain.PathFinder;
 import nextstep.subway.station.application.StationService;
 import nextstep.subway.station.domain.Station;
 import org.springframework.stereotype.Service;
