+++ src/main/java/nextstep/subway/path/domain/DijkstraPathFinder.java
@@ -1,4 +1,4 @@
-package nextstep.subway.path.finder;
+package nextstep.subway.path.domain;
 
 import java.util.Collection;
 import java.util.List;
